[#1323] lic.equinox plug-in to be renamed
* reworked dependencies to equinox
* moved helper types to LIC Execute
* extracted LIC CLI bundle

Fixes #1323
