+++ bundles/org.eclipse.passage.lbc.base/META-INF/MANIFEST.MF
@@ -7,10 +7,10 @@ Bundle-Name: %Bundle-Name
 Bundle-Vendor: %Bundle-Vendor
 Bundle-Copyright: %Bundle-Copyright
 Bundle-RequiredExecutionEnvironment: JavaSE-17
-Require-Bundle: org.eclipse.passage.lic.base;bundle-version="0.0.0",
+Require-Bundle: org.eclipse.emf.ecore.xmi;bundle-version="0.0.0",
+ org.eclipse.passage.lic.base;bundle-version="0.0.0",
  org.eclipse.passage.lic.net;bundle-version="0.0.0",
  org.eclipse.passage.lic.licenses.model;bundle-version="0.0.0",
- org.eclipse.emf.ecore.xmi;bundle-version="0.0.0",
  org.eclipse.passage.lic.emf;bundle-version="0.0.0",
  org.eclipse.passage.lic.oshi;bundle-version="0.0.0",
  org.eclipse.passage.lic.equinox;bundle-version="0.0.0"
