+++ bundles/org.eclipse.passage.lbc.fls.seal/src/org/eclipse/passage/lbc/internal/fls/seal/SealedAccessCycleConfiguration.java
@@ -27,9 +27,9 @@ import org.eclipse.passage.lic.base.conditions.mining.UserHomeResidentConditions
 import org.eclipse.passage.lic.base.registry.ReadOnlyRegistry;
 import org.eclipse.passage.lic.equinox.acquire.ConfigurationLicenseAcquisitionService;
 import org.eclipse.passage.lic.equinox.acquire.InstallationLicenseAcquisitionService;
-import org.eclipse.passage.lic.equinox.conditions.ConfigurationResidentConditions;
-import org.eclipse.passage.lic.equinox.conditions.InstallationResidentConditions;
 import org.eclipse.passage.lic.execute.BaseAccessCycleConfiguration;
+import org.eclipse.passage.lic.execute.ConfigurationResidentConditions;
+import org.eclipse.passage.lic.execute.InstallationResidentConditions;
 import org.osgi.framework.FrameworkUtil;
 
 final class SealedAccessCycleConfiguration extends BaseAccessCycleConfiguration {
