+++ bundles/org.eclipse.passage.lic.cli/META-INF/MANIFEST.MF
@@ -1,0 +1,14 @@
+Manifest-Version: 1.0
+Automatic-Module-Name: org.eclipse.passage.lic.cli
+Bundle-ManifestVersion: 2
+Bundle-SymbolicName: org.eclipse.passage.lic.cli
+Bundle-Version: 3.0.0.qualifier
+Bundle-Name: %Bundle-Name
+Bundle-Vendor: %Bundle-Vendor
+Bundle-Copyright: %Bundle-Copyright
+Bundle-RequiredExecutionEnvironment: JavaSE-17
+Require-Bundle: org.eclipse.core.runtime;bundle-version="0.0.0";visibility:=reexport,
+ org.eclipse.passage.lic.equinox;bundle-version="0.0.0";visibility:=reexport
+Export-Package: org.eclipse.passage.lic.cli;x-friends:="org.eclipse.passage.lic.jetty"
+Import-Package: org.slf4j;version="1.7.36"
+Bundle-ActivationPolicy: lazy
