+++ bundles/org.eclipse.passage.lic.cli/OSGI-INF/l10n/bundle.properties
@@ -1,0 +1,22 @@
+###############################################################################
+# Copyright (c) 2018, 2024 ArSysOp and others
+#
+# This program and the accompanying materials are made available under the
+# terms of the Eclipse Public License 2.0 which is available at
+# https://www.eclipse.org/legal/epl-2.0/.
+#
+# SPDX-License-Identifier: EPL-2.0
+#
+# Contributors:
+#     ArSysOp - initial API and implementation
+###############################################################################
+
+Bundle-Name = Passage LIC CLI
+Bundle-Vendor = Eclipse Passage
+Bundle-Copyright = Copyright (c) 2018, 2024 ArSysOp and others.\n\
+\n\
+This program and the accompanying materials are made\n\
+available under the terms of the Eclipse Public License 2.0\n\
+which is available at https://www.eclipse.org/legal/epl-2.0/\n\
+\n\
+SPDX-License-Identifier: EPL-2.0\n\
