+++ bundles/org.eclipse.passage.lic.cli/src/org/eclipse/passage/lic/cli/LicenseCoverageCheck.java
@@ -10,7 +10,7 @@
  * Contributors:
  *     ArSysOp - initial API and implementation
  *******************************************************************************/
-package org.eclipse.passage.lic.equinox.access;
+package org.eclipse.passage.lic.cli;
 
 import java.util.ArrayList;
 import java.util.Collection;
