+++ bundles/org.eclipse.passage.lic.cli/src/org/eclipse/passage/lic/cli/OptionDiagnostic.java
@@ -10,7 +10,7 @@
  * Contributors:
  *     ArSysOp - initial API and implementation
  *******************************************************************************/
-package org.eclipse.passage.lic.equinox.access;
+package org.eclipse.passage.lic.cli;
 
 import org.eclipse.passage.lic.api.diagnostic.Diagnostic;
 import org.eclipse.passage.lic.base.diagnostic.DiagnosticExplained;
