+++ bundles/org.eclipse.passage.lic.equinox/META-INF/MANIFEST.MF
@@ -7,11 +7,12 @@ Bundle-Name: %Bundle-Name
 Bundle-Vendor: %Bundle-Vendor
 Bundle-Copyright: %Bundle-Copyright
 Bundle-RequiredExecutionEnvironment: JavaSE-17
-Require-Bundle: org.eclipse.core.runtime;bundle-version="0.0.0";visibility:=reexport,
+Require-Bundle: org.eclipse.equinox.app;bundle-version="0.0.0";visibility:=reexport,
+ org.eclipse.equinox.registry;bundle-version="0.0.0",
+ org.eclipse.osgi;bundle-version="0.0.0";visibility:=reexport,
  org.eclipse.passage.lic.api;bundle-version="0.0.0";visibility:=reexport,
  org.eclipse.passage.lic.base;bundle-version="0.0.0";visibility:=reexport
 Export-Package: org.eclipse.passage.lic.equinox;x-internal:=true,
- org.eclipse.passage.lic.equinox.access;x-friends:="org.eclipse.passage.lic.jetty",
  org.eclipse.passage.lic.equinox.acquire;x-internal:=true,
  org.eclipse.passage.lic.equinox.conditions;x-internal:=true,
  org.eclipse.passage.lic.equinox.io;x-internal:=true,
