+++ bundles/org.eclipse.passage.lic.equinox/src/org/eclipse/passage/lic/internal/equinox/i18n/EquinoxMessages.properties
@@ -1,5 +1,5 @@
 ###############################################################################
-# Copyright (c) 2019, 2021 ArSysOp and others
+# Copyright (c) 2019, 2024 ArSysOp and others
 #
 # This program and the accompanying materials are made available under the
 # terms of the Eclipse Public License 2.0 which is available at
@@ -20,6 +20,5 @@ BundleVendor_unknown_vendor=Unknown vendor
 LicensedApplication_application_context_service_unregistered=Severe runtime insufficiency error: IApplicationContext service is unregistered or unavailable
 LicensedApplication_no_application_context_service_ref=Severe runtime insufficiency error: no service reference for IApplicationContext
 LicensedApplication_no_bundle_context=Severe runtime insufficiency error: no bundle context
-ProductContacts_defaults=Protected by Eclipse Passage. Please visit https://eclipse-passage.github.io/passage-docs/#support
 RequirementsFromCapability_no_attributes=Attributes for capability {0} in bundle {1} is required
 RequirementsFromCapability_no_feature_id=Configuration of a feature identifier in capability {0} of bundle {1} is required
