+++ bundles/org.eclipse.passage.lic.jetty/src/org/eclipse/passage/lic/internal/jetty/interaction/ServerHandles.java
@@ -15,7 +15,7 @@ package org.eclipse.passage.lic.internal.jetty.interaction;
 import java.util.Arrays;
 import java.util.List;
 
-import org.eclipse.passage.lic.equinox.access.LicenseProtection;
+import org.eclipse.passage.lic.cli.LicenseProtection;
 import org.eclipse.passage.lic.internal.jetty.JettyException;
 import org.eclipse.passage.lic.internal.jetty.JettyServer;
 import org.eclipse.passage.lic.internal.net.connect.BindAddress;
