+++ bundles/org.eclipse.passage.lic.jface/META-INF/MANIFEST.MF
@@ -7,7 +7,8 @@ Bundle-Name: %Bundle-Name
 Bundle-Vendor: %Bundle-Vendor
 Bundle-Copyright: %Bundle-Copyright
 Bundle-RequiredExecutionEnvironment: JavaSE-17
-Require-Bundle: org.eclipse.equinox.common;bundle-version="0.0.0";visibility:=reexport,
+Require-Bundle: org.eclipse.core.runtime;bundle-version="0.0.0",
+ org.eclipse.equinox.common;bundle-version="0.0.0";visibility:=reexport,
  org.eclipse.jface;bundle-version="0.0.0";visibility:=reexport,
  org.eclipse.passage.lic.api;bundle-version="0.0.0";visibility:=reexport,
  org.eclipse.passage.lic.base;bundle-version="0.0.0";visibility:=reexport,
