+++ bundles/org.eclipse.passage.lic.jface/src/org/eclipse/passage/lic/internal/jface/dialogs/licensing/BaseLicenseStatusDialog.java
@@ -24,7 +24,6 @@ import org.eclipse.passage.lic.base.diagnostic.NoErrors;
 import org.eclipse.passage.lic.base.diagnostic.RequirementStatus;
 import org.eclipse.passage.lic.base.diagnostic.RequirementsCoverage;
 import org.eclipse.passage.lic.base.restrictions.ExaminationExplained;
-import org.eclipse.passage.lic.equinox.ProductContacts;
 import org.eclipse.passage.lic.internal.jface.i18n.LicenseStatusDialogMessages;
 import org.eclipse.swt.SWT;
 import org.eclipse.swt.custom.StyledText;
