+++ bundles/org.eclipse.passage.lic.jface/src/org/eclipse/passage/lic/internal/jface/dialogs/licensing/ProductContacts.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020, 2021 ArSysOp
+ * Copyright (c) 2020, 2024 ArSysOp
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
@@ -10,13 +10,13 @@
  * Contributors:
  *     ArSysOp - initial API and implementation
  *******************************************************************************/
-package org.eclipse.passage.lic.equinox;
+package org.eclipse.passage.lic.internal.jface.dialogs.licensing;
 
 import java.util.Optional;
 import java.util.function.Supplier;
 
 import org.eclipse.core.runtime.Platform;
-import org.eclipse.passage.lic.internal.equinox.i18n.EquinoxMessages;
+import org.eclipse.passage.lic.internal.jface.i18n.LicenseStatusDialogMessages;
 
 /**
  * @since 2.1
@@ -33,7 +33,7 @@ public final class ProductContacts implements Supplier<String> {
 	}
 
 	public String defaults() {
-		return EquinoxMessages.ProductContacts_defaults;
+		return LicenseStatusDialogMessages.ProductContacts_defaults;
 	}
 
 }
