+++ bundles/org.eclipse.passage.lic.jface/src/org/eclipse/passage/lic/internal/jface/i18n/LicenseStatusDialogMessages.properties
@@ -1,5 +1,5 @@
 ###############################################################################
-# Copyright (c) 2020, 2021 ArSysOp and others
+# Copyright (c) 2020, 2024 ArSysOp and others
 #
 # This program and the accompanying materials are made available under the
 # terms of the Eclipse Public License 2.0 which is available at
@@ -23,6 +23,7 @@ LicenseStatusDialog_intention_accept_tooltip=There are not yet accepted license 
 LicenseStatusDialog_intention_diagnose=&Diagnostic
 LicenseStatusDialog_intention_diagnose_tooltip=There is a diagnostic information available, view it in a dedicated dialog
 LicenseStatusDialog_title=Licensing status
+ProductContacts_defaults=Protected by Eclipse Passage. Please visit https://eclipse-passage.github.io/passage-docs/#support
 WithCertificatSummary_error=License coverage is not sufficient and has restrictive statements
 WithCertificatSummary_ok=License coverage is sufficient
 WithCertificatSummary_warning=License coverage is not sufficient and demands attention
