+++ bundles/org.eclipse.passage.loc.workspace/META-INF/MANIFEST.MF
@@ -11,5 +11,6 @@ Bundle-ActivationPolicy: lazy
 Require-Bundle: org.eclipse.core.runtime;bundle-version="0.0.0",
  org.eclipse.passage.lic.base;bundle-version="0.0.0",
  org.eclipse.passage.lic.equinox;bundle-version="0.0.0",
- org.eclipse.passage.loc.api;bundle-version="0.0.0"
+ org.eclipse.passage.loc.api;bundle-version="0.0.0",
+ org.eclipse.passage.lic.execute;bundle-version="3.0.0"
 Export-Package: org.eclipse.passage.loc.workspace;x-friends:="org.eclipse.passage.loc.operator.gear"
