+++ bundles/org.eclipse.passage.loc.workspace/src/org/eclipse/passage/loc/workspace/UncertainProducts.java
@@ -2,7 +2,7 @@ package org.eclipse.passage.loc.workspace;
 
 import java.util.List;
 
-import org.eclipse.passage.lic.equinox.io.InstallationPath;
+import org.eclipse.passage.lic.execute.InstallationPath;
 import org.eclipse.passage.loc.internal.api.workspace.Products;
 import org.eclipse.passage.loc.internal.api.workspace.ResourceHandle;
