+++ features/org.eclipse.passage.lic.equinox.feature/feature.xml
@@ -47,6 +47,13 @@
          unpack="false"/>
 
    <plugin
+         id="org.eclipse.passage.lic.cli"
+         download-size="0"
+         install-size="0"
+         version="0.0.0"
+         unpack="false"/>
+
+   <plugin
          id="org.eclipse.passage.lic.equinox"
          download-size="0"
          install-size="0"
