+++ releng/org.eclipse.passage.lic.aggregator/pom.xml
@@ -30,11 +30,11 @@
 
 	<modules>
 		<module>../../bundles/org.eclipse.passage.lic.api</module>
-		<module>../../tests/org.eclipse.passage.lic.api.tests</module>
-
 		<module>../../bundles/org.eclipse.passage.lic.base</module>
+		<module>../../bundles/org.eclipse.passage.lic.cli</module>
 		<module>../../bundles/org.eclipse.passage.lic.equinox</module>
 		<module>../../features/org.eclipse.passage.lic.equinox.feature</module>
+		<module>../../tests/org.eclipse.passage.lic.api.tests</module>
 		<module>../../tests/org.eclipse.passage.lic.base.tests</module>
 		<module>../../tests/org.eclipse.passage.lic.equinox.tests</module>
 		<module>../../tests/org.eclipse.passage.lic.equinox.tests.data.requirements</module>
