+++ tests/org.eclipse.passage.lic.equinox.tests/META-INF/MANIFEST.MF
@@ -9,7 +9,7 @@ Bundle-Vendor: %Bundle-Vendor
 Bundle-Copyright: %Bundle-Copyright
 Fragment-Host: org.eclipse.passage.lic.equinox
 Import-Package: org.junit
-Require-Bundle: org.eclipse.equinox.event;bundle-version="0.0.0",
+Require-Bundle: org.eclipse.core.runtime;bundle-version="0.0.0",
  org.eclipse.passage.lic.base;bundle-version="0.0.0",
  org.eclipse.passage.lic.equinox.tests.data.requirements;bundle-version="0.1.0",
  org.eclipse.passage.lic.base.tests;bundle-version="1.0.0",
